Изменен способ парсинга настроек процессоров Добавлено форматирование времени в процессор таймер Добавлено наложение маски в процессор изображение Добавлена возможность наложения глитча Рефакторинг для наложения глитчей

+++ internal/transformer/frameTransformer.go
@@ -1,9 +1,9 @@
 package transformer
 
-func SecondsToFrameCount(seconds float32, framesPerSecond int) int {
+func SecondsToFrameCount(seconds float64, framesPerSecond int) int {
 	return int(seconds) * framesPerSecond
 }
 
-func FrameCountToSeconds(frameCount int, framesPerSecond int) float32 {
-	return float32(frameCount / framesPerSecond)
+func FrameCountToSeconds(frameCount int, framesPerSecond int) float64 {
+	return float64(frameCount) / float64(framesPerSecond)
 }
